% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.r
\docType{data}
\name{deg}
\alias{deg}
\title{A Gene List Data to Test Functions.

Data of genes with hgnc symbol and entrez gene id.

\itemize{
    \item hgnc_symbol. hgnc gene symbol.
    \item entrezgene. entrez gene id.
}}
\format{a data.frame of genes, with columns hgnc_symbol and entrezgene.}
\usage{
deg
}
\keyword{datasets}

