% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGGPathway.r
\name{enrichKEGGPathway}
\alias{enrichKEGGPathway}
\title{KEGG pathway enrichment tests.

\code{enrichKEGGPathway} will do the hypergeometric tests to get the p value and
adjusted p value of KEGG pathways.}
\usage{
enrichKEGGPathway(entrez.gene.list, pathway.list = NA,
  whole.gene.number = 22333, threshold = 0.05, adjust.p.method = "BH",
  organism = "hsa")
}
\arguments{
\item{entrez.gene.list}{character vector, entrez gene IDs of DE genes.,}

\item{pathway.list}{character vector, KEGG pathway IDs.
If NA, the function will use all the pathways in KEGG website automatically.}

\item{whole.gene.number}{numeric, the total backgroud number of all genes.}

\item{threshold}{numeric, output tests' adjusted p value cut off, default 0.05.}

\item{adjust.p.method}{character, method used in \code{p.adjust}.}

\item{organism}{character, 3 characters' KEGG organism code, default hsa (human).}
}
\value{
data.frame with columns,
Term (pathway names),
Type (always KEGG pathway),
ID (KEGG pathway ID),
Input.number (number of input genes in the pathway),
Background.number (number of all genes in the pathway),
P.value,
adjusted.P.value.
}
\details{
\code{enrichKEGGPathway} is used for pathway enrichment.
The pathway information is fetched from KEGG website.
\code{phyper} is used to do the hypergeometric tests.
\code{p.adjust} is used to calculate the p adjustes.
}
\examples{
data("deg")

de.entrez <- as.character(na.omit(deg$entrezgene))

enrich.result <- enrichKEGGPathway(
    entrez.gene.list  = de.entrez,
    pathway.list      = NA,
    whole.gene.number = 20000,
    threshold         = 0.05,
    adjust.p.method   = "BH",
    organism          = "hsa"
)

enrich.result

}
\seealso{
\code{\link{viewKEGGPathwayInfo}}
\code{\link{p.adjust}}
\code{\link{phyper}}
}

