% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGGPathway.r
\name{enrichKEGGPathwayParent}
\alias{enrichKEGGPathwayParent}
\title{KEGG pathway enrichment tests

The function to do hypergeometric tests on KEGG pathways.}
\usage{
enrichKEGGPathwayParent()
}
\value{
the returned result will be a data.frame,
containing numbers of drawn, background number, p value and adjusted p value,
with the same row number as the length of input \code{urns}.
The order of results will be the same with \code{urns},
and the names will also be the same if has.
}
\details{
This function is used to do many hypergeometric tests,
and calculate both the p value and adjusted p value.
}
\seealso{
\code{\link{enrichKEGGPathway}}
}

