% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMSigDB.r
\name{enrichMSigDB}
\alias{enrichMSigDB}
\title{MSigDB data sets enrichment tests.

\code{enrichMSigDB} will do the hypergeometric tests to get the p value and
adjusted p value of KEGG pathways.}
\usage{
enrichMSigDB(entrez.gene.list, msigdb = NA, whole.gene.number = 22333,
  threshold = 0.05, adjust.p.method = "BH")
}
\arguments{
\item{entrez.gene.list}{character vector, entrez gene IDs of DE genes.,}

\item{msigdb}{list of MSigDBSet.
If NA, the function will load data MSigDB of the package automatically.}

\item{whole.gene.number}{numeric, the total backgroud number of all genes.}

\item{threshold}{numeric, output tests' adjusted p value cut off, default 0.05.}

\item{adjust.p.method}{character, method used in \code{p.adjust}.}
}
\value{
data.frame with columns,
Term (MSigDB set names),
Type (always MSigDB),
ID (MSigDB set SYSTEMATIC_NAME),
Input.number (number of input genes in the set),
Background.number (number of all genes in the set),
P.value,
adjusted.P.value.
}
\details{
\code{enrichMSigDB} is used for MSigDB sets enrichment.
The sets come from MSigDB web
(\url{http://software.broadinstitute.org/gsea/msigdb}).
}
\examples{
data("deg")

de.entrez <- as.character(na.omit(deg$entrezgene))

enrich.result <- enrichMSigDB(
    entrez.gene.list  = de.entrez,
    msigdb            = NA,
    whole.gene.number = 20000,
    threshold         = 0.05,
    adjust.p.method   = "BH"
)

enrich.result

}
\seealso{
\code{\link{enrichKEGGPathway}}
\code{\link{MSigDB}}
\code{\link{MSigDBSet}}
}

