% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPathwayInfo.r
\name{fetchPathwayInfo}
\alias{fetchPathwayInfo}
\title{Fetch single pathway information from KEGG website

\code{.hypergemetricTestVector} helps calculate the p value and adjusted p value
in the hypergeometric tests of intersection of one set to defined sets.}
\usage{
fetchPathwayInfo(pathway.id)
}
\arguments{
\item{pathway.id}{kegg id  of pathway, length should be 1.}
}
\value{
the returned result will be a list,
contain the information of certain pathway fetched from KEGG website API.
}
\details{
This function is used to do many hypergeometric tests,
and calculate both the p value and adjusted p value.
}
\seealso{
\code{\link{phyper}} \code{\link{p.adjust}}
}

