% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.project.id.R
\name{find.project.id}
\alias{find.project.id}
\title{Any rows with duplicated row names will be dropped with the first one being
kepted.}
\usage{
find.project.id(df, title, providers_id, project_privacy)
}
\arguments{
\item{df}{result df from index.all.projects}

\item{title}{of project or matching term}

\item{providers_id}{of project or matching term}

\item{project_privacy}{of project or matching term}
}
\value{
a subsetted dataframe of index.all.projects which includes a column with relevant project unique_ids that can be passed to, e.g., find.projects
}
\description{
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
