% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.projects.R
\name{find.projects}
\alias{find.projects}
\title{Pull data on specific project(s) by their unique project id}
\usage{
find.projects(proj)
}
\arguments{
\item{proj}{project id or list of project ids}
}
\value{
A matrix of the infile
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kept.
}
