% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometricTestVector.r
\name{hypergeometricTestVector}
\alias{hypergeometricTestVector}
\title{Hypergeomatric Test of Terms and Genes

\code{.hypergemetricTestVector} helps calculate the p value and adjusted p value
in the hypergeometric tests of intersection of one set to defined sets.}
\usage{
hypergeometricTestVector(drawn.ball, urns, total.ball.num,
  adj.p.method = "BH")
}
\arguments{
\item{drawn.ball}{vector of items to be tested, like drawn balls.}

\item{urns}{lists of items contained in different sets, like different urns
contains balls.}

\item{total.ball.num}{the background number of items.}

\item{adj.p.method}{adjusted p value methods in \code{p.adjust}, including
\code{c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")},
default to be "BH".}
}
\value{
the returned result will be a data.frame,
containing numbers of drawn, background number, p value and adjusted p value,
with the same row number as the length of input \code{urns}.
The order of results will be the same with \code{urns},
and the names will also be the same if has.
}
\details{
This function is used to do many hypergeometric tests,
and calculate both the p value and adjusted p value.
}
\seealso{
\code{\link{phyper}} \code{\link{p.adjust}}
}

