% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.all.projects.R
\name{index.all.projects}
\alias{index.all.projects}
\title{Load all Hub projects}
\usage{
index.all.projects()
}
\value{
all project in TK Hub
}
\description{
This function loads a dataframe of all projects and their unique ids
As the Hub grows and/or internet stability this can take a while to run
}
