% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKEGGOrganism.r
\name{listKEGGOrganism}
\alias{listKEGGOrganism}
\title{List available organisms of KEGG website.

\code{listKEGGOrganism} returns a data.frame contains the information about
available organisms in KEGG website.}
\usage{
listKEGGOrganism(origin = FALSE)
}
\arguments{
\item{origin}{bool, if TRUE, return the file fetched from KEGG website.
If FALSE, the file contents will be reordered.}
}
\value{
the returned result will be a data.frame, containing the information of
organisms in KEGG website. The data.frame has organisms' ID,
organisms' three characters code, and organisms' scientific and English names.
}
\details{
\code{listKEGGOrganism} using API of KEGG website
(\url{"http://rest.kegg.jp/list/organism"}) to fetch the lists of organisms.
The three-character KEGG organism code is used as the start 3 characters
of KEGG pathway ID.
}
\examples{
organism <- listKEGGOrganism()

organism

}
\seealso{
\code{\link{listKEGGPathway}}
}

