% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKEGGPathway.r
\name{listKEGGPathway}
\alias{listKEGGPathway}
\title{List available pathways of KEGG website.

\code{listKEGGPathway} returns a data.frame contains the information about
pathways for certain organism in KEGG website.}
\usage{
listKEGGPathway(organism, origin = FALSE)
}
\arguments{
\item{organism}{character, 3 characters KEGG organism code.}

\item{origin}{bool, if TRUE, return the file fetched from KEGG website.
If FALSE, the file contents will be reordered.}
}
\value{
data.frame with KEGG pathway IDs and names.
}
\details{
\code{listKEGGOrganism} using API of KEGG website
(\url{"http://rest.kegg.jp/list/pathway/"}) to fetch the lists of available
pathways of certain organism. The parameter, organism, could be KEGG organism
code, normal English name, scientific name. Organisms in KEGG website could be
abtained from \code{listKEGGOrganism}.
}
\examples{
pathways <- listKEGGPathway()

pathways

}
\seealso{
\code{\link{listKEGGOrganism}}
}

