% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigDB.r
\docType{data}
\name{MSigDB}
\alias{MSigDB}
\title{MSigDB data

Data of the MSigDB sets, stored in list. using \code{data("msigdb")} to load data.}
\format{a list with 13311 objects of class "MSigDBSet".}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb}
}
\usage{
MSigDB
}
\keyword{datasets}

