% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigDBSet.r
\docType{class}
\name{MSigDBSet-class}
\alias{MSigDBSet-class}
\alias{getMembers,MSigDBSet-method}
\alias{initialize,MSigDBSet-method}
\title{MSigDB data set class

The Class "MSigDBSet" is used to store set from MSigDB.
The original data is downloaded from MSigDB webset with XML format.}
\usage{
\S4method{initialize}{MSigDBSet}(.Object, value.vector = NULL,
  STANDARD_NAME = "", SYSTEMATIC_NAME = "", HISTORICAL_NAMES = "",
  ORGANISM = "", PMID = "", AUTHORS = "", GEOID = "",
  EXACT_SOURCE = "", GENESET_LISTING_URL = "", EXTERNAL_DETAILS_URL = "",
  CHIP = "", CATEGORY_CODE = "", SUB_CATEGORY_CODE = "",
  CONTRIBUTOR = "", CONTRIBUTOR_ORG = "", DESCRIPTION_BRIEF = "",
  DESCRIPTION_FULL = "", TAGS = "", MEMBERS = "",
  MEMBERS_SYMBOLIZED = "", MEMBERS_EZID = "", MEMBERS_MAPPING = "",
  FOUNDER_NAMES = "", REFINEMENT_DATASETS = "", VALIDATION_DATASETS = "",
  ...)

\S4method{getMembers}{MSigDBSet}(.Object, type = c("member", "symbol",
  "entrez"), ...)
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: initialize the MSigDBSet object.

\item \code{getMembers}: return slot values of MSigDBSet object.
}}
\section{Slots}{

\describe{
\item{\code{STANDARD_NAME}}{character, standard name of the set recently.}

\item{\code{SYSTEMATIC_NAME}}{character, systematic name of the sets,
unique and rarely changed.}

\item{\code{HISTORICAL_NAMES}}{character, names of the set used in the past.}

\item{\code{ORGANISM}}{character, organism of the set.}

\item{\code{PMID}}{character, PMID of the set.}

\item{\code{AUTHORS}}{character.}

\item{\code{GEOID}}{character.}

\item{\code{EXACT_SOURCE}}{character.}

\item{\code{GENESET_LISTING_URL}}{character, url of the set in MSigDB.}

\item{\code{EXTERNAL_DETAILS_URL}}{character, url of the set in MSigDB.}

\item{\code{CHIP}}{character.}

\item{\code{CATEGORY_CODE}}{character, category code of set.
H: hallmark gene sets. 
C1: positional gene sets. 
C2: curated gene sets, with subcategories (CGP, CP, CP:BIOCARTA, CP:KEGG, CP:REACTOME). 
C3: motif gene sets, with subcategories (MIR, TFT). 
C4: computational gene sets, with subcategories (CGN, CM). 
C5: GO gene sets, with subcategories (BP, CC, MF).
C6: oncogenic signatures.
C7: immunologic signatures.}

\item{\code{SUB_CATEGORY_CODE}}{character, sub category code of set. 
C2: 
CGP (chemical and genetic perturbations), 
CP (Canonical pathways), 
CP:BIOCARTA (BioCarta gene sets), 
CP:KEGG (KEGG gene sets), 
CP:REACTOME (Reactome gene sets). 
C3:
MIR (microRNA targets), 
TFT (transcription factor targets).
C4:
CGN (cancer gene neighborhoods),
CM (cancer modules).
C5: 
BP (GO biological process),
CC (GO cellular component).}

\item{\code{CONTRIBUTOR}}{character.}

\item{\code{CONTRIBUTOR_ORG}}{character.}

\item{\code{DESCRIPTION_BRIEF}}{character, brif description of the set.}

\item{\code{DESCRIPTION_FULL}}{character, detailed description of the set.}

\item{\code{TAGS}}{character.}

\item{\code{MEMBERS}}{character, gene members with label of MSigDB.}

\item{\code{MEMBERS_SYMBOLIZED}}{character, gene members with gene symbol.}

\item{\code{MEMBERS_EZID}}{character, gene members with entrez id.}

\item{\code{MEMBERS_MAPPING}}{character, gene members with mapping of three labels.}

\item{\code{FOUNDER_NAMES}}{character.}

\item{\code{REFINEMENT_DATASETS}}{character.}

\item{\code{VALIDATION_DATASETS}}{character.}
}}
\examples{

data("msigdb")

genes <- getMembers(
    MSigDB[[1]],
    type = "symbol"
)

genes 

}

