% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputFASTA.R
\name{outputFASTA}
\alias{outputFASTA}
\title{Append Local Contexts Unique ID to FASTA headers}
\usage{
outputFASTA(seqs, seqid, uqID, filename)
}
\arguments{
\item{seqs}{sequences of interest}

\item{seqid}{original sequence headers}

\item{uqID}{Local Contexts Hub Unique ID}

\item{filename}{desired FASTA file name (no .fasta needed)}
}
\value{
A FASTA file with HUB unique ID appended to headers
}
\description{
This function writes a FASTA file with Local Contexts Hub Unique ID appended to sequence headers
}
