% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFASTA.R
\name{readFASTA}
\alias{readFASTA}
\title{import FASTA files}
\usage{
readFASTA(file)
}
\arguments{
\item{file}{the name of, and/or path to, your FASTA}
}
\value{
a dataframe of seq headers and associated seq
}
\description{
This function allows you to import FASTA files into R as a dataframe
}
