% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdbxml.r
\name{readMSigDBxml}
\alias{makeMSigDBList}
\alias{readMSigDBxml}
\title{MSigDB XML file processing

These functions are used to deal with the xml file downloaded from MSigDB.}
\usage{
readMSigDBxml(file)

makeMSigDBList(file)
}
\arguments{
\item{file}{xml file downloaded from MSigDB (\url{http://software.broadinstitute.org/gsea/msigdb}).}
}
\value{
a list of processed XML or list of "MSigDBSet" objects.
}
\details{
\code{readMSigDBxml} is a wrapped function of xmlToList.
\code{makeMSigDBList} returns a list of "MSigDBSet" objects.
}

