% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_window.R
\name{slide_window}
\alias{slide_window}
\title{Slide window}
\usage{
slide_window(sequence, window_size, slide_size)
}
\description{
This function will create kmers by sliding a window in the fragment. The window size
is 120 bp and the slide_size can be changes as you need. slide_size=1 indicates that a new 120-kmer
will be generated at each nucleotide.
}
