% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigDB.r
\name{subsetMSigDB}
\alias{subsetMSigDB}
\title{subset MSigDB data

subsetMSigDB is used to subset the MSigDB data (list of "MSigDBSet" objects)
with slot names and values of "MSigDBSet" in MSigDB data.}
\usage{
subsetMSigDB(msigdb, slot.type, slot.value)
}
\arguments{
\item{msigdb}{the list of "MSigDBSet" objects.}

\item{slot.type}{the name of slot of "MSigDBSet".}

\item{slot.value}{the value of slot selected.}
}
\value{
a sublist of the data inputed that match the value gaven.
}
\details{

}
\examples{

data("msigdb")

mysublist <- subsetMSigDB(
    msigdb     = MSigDB,
    slot.type  = "ORGANISM",
    slot.value = "Homo sapiens"
)
    

}
\seealso{
\code{\link{MSigDB}}
}

