% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFASTA.R
\name{testFASTA}
\alias{testFASTA}
\title{Generate test FASTAs or sequences}
\usage{
testFASTA(file_name, seqlength, DNA_prob, no_seqs, out.fasta, loc)
}
\arguments{
\item{file_name}{name of FASTA file to write if out.fasta is TRUE}

\item{seqlength}{vector of length 2 specifying the min and max seq length range}

\item{DNA_prob}{a vector of length 4 specifying the probabilities of nucleotides A, T, G, and C}

\item{no_seqs}{the number of sequences to simulate}

\item{out.fasta}{whether to write a FASTA file or output a df of seqs}

\item{loc}{name of locus}
}
\value{
A matrix of the infile
}
\description{
If out.fasta is TRUE a dummy fasta file will be written
If out.fasta is FALSE a test df of seqs will be created
}
