% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewKEGGPathwayInfo.r
\name{viewKEGGPathwayInfo}
\alias{viewKEGGPathwayInfo}
\title{view pathway information of KEGG website.

\code{viewKEGGPathwayInfo} returns information about KEGG pathway fetched from
KEGG website.}
\usage{
viewKEGGPathwayInfo(pathway.id)
}
\arguments{
\item{pathway.id}{character, KEGG pathway ID,
namely with three characters and several numbers, such as map01100.}
}
\value{
list of KEGG pathways information.
If the pathway.id only have one pathway ID,
the result is a list contain the information of that pathway.
If the pathway.id contains many pathway,
the result is a list contains many pathway information lists.
}
\details{
\code{viewKEEGGPathwayInfo} will return the information about a certain
pathway fetched from KEGG website.
The result is a list contains the information of one or more pathways.
The function just fetchs the information of pathway from KEGG website,
using KEGG website API \url("http://rest.kegg.jp/get/map01100").
}
\examples{
map01100 <- viewKEGGPathwayInfo("map01100")

map01100

}
\seealso{
\code{\link{listKEGGPathway}} \code{\link{listKEGGOrganism}}
}

