\name{enRich-package}
\alias{enRich-package}
\alias{enRich}
\docType{package}
\title{
An R package for analysis of multiple ChIP-seq data.
}
\description{
enRich is an R package that performs a joint statistical modelling of ChIP-seq data, accounting for technical/biological replicates, multiple conditions and
the different IP efficiencies of individual experiments. 
}
\details{
\tabular{ll}{
Package: \tab enRich\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-22\cr
Depends: \tab R(>= 2.14), parallel\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
mainfunctions: mix, mix.joint, enrich.mix
}
\author{
Yanchun Bao \email{yanchun.bao@brunel.ac.uk} and 
\cr Veronica Vinciotti \email{veronica.vinciotti@brunel.ac.uk}

Maintainer: Yanchun Bao \email{yanchun.bao@brunel.ac.uk} and
\cr Veronica Vinciotti \email{veronica.vinciotti@brunel.ac.uk}
}
\references{
Bao et al. (2013). Accounting for immunoprecipitation efficiencies in the statistical analysis of ChIP-seq data. To appear in BMC Bioinformatics.
}
\keyword{ package}


