% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\title{enRich
This is the main function of the package.}
\usage{
enrich(
  ref,
  bed_file,
  frag_size,
  w_size,
  s_size,
  gc_min,
  gc_max,
  blast_db,
  makeblastdb_path,
  blastn_path,
  blast_out,
  pid,
  len,
  blast_hit,
  tm_min,
  tm_max,
  HpTm,
  HmTm,
  threads
)
}
\description{
enRich
This is the main function of the package.
}
\examples{
# example code
#input values
ref= system.file("extdata", "genome.fasta", package = "enRich")# reference genome
file.copy(ref,"./", overwrite = TRUE) #saving reference genome in the directory
index=system.file("extdata", "genome.fasta.fai", package = "enRich")# reference genome index
file.copy(index,"./", overwrite = TRUE) #saving index in the current directory
bed_file= system.file("extdata", "bed_test.bed", package = "enRich") # bed file
file.copy(bed_file,"./", overwrite = TRUE) #savind bed file in the current directory
blast_db= "./db/genome" #blast database name
makeblastdb_path= "/path/to//blast/ncbi-blast-2.1X.0+/bin/makeblastdb" 
blastn_path= "/path/to/blast/ncbi-blast-2.1X.0+/bin/blastn" 
blast_out="./blast_result.txt"
frag_size=400 # frament size in bp for the region with the target SNP in the center
w_size= 120 #size of the probe
s_size= 1 #number of oligonucleotides to move
gc_min= 40 # minimum GC\%
gc_max= 46 # maximum GC\%
blast_db= "./db/genome" #blast database name
pid= 90 # percentage identity of blast result (default 90)
len= 120 # length of the match in the blast result (default 120)
blast_hit= 1 # maximum number of blast match for the probe (default 1)
tm_min= 70 # minimum Tm (default 70)
tm_max= 76 #maximum Tm (default 76)
HpTm= 60 #Hairpin maximum Tm (default 60)
HmTm= 60 # Heterodimer maximum Tm (default 60)
threads= 16 # set the number based on the request from the cluster
# Run the function
library(enRich)
enrich("genome.fasta", "bed_test.bed", frag_size, w_size, s_size, gc_min, gc_max, blast_db, pid, len, blast_hit,tm_min, tm_max, HpTm, HmTm, threads)
}
